+++ src/presentation/controllers/login/login-protocols.ts
@@ -1,3 +1,3 @@
 export * from '../../protocols'
 export * from '../../../domain/usecases/authenticaton'
-export * from '../../helpers/validators/validation'
+export * from '../../protocols/validation'
