+++ src/presentation/helpers/validators/validation-composite.ts
@@ -1,4 +1,4 @@
-import { type Validation } from './validation'
+import { type Validation } from '../../protocols/validation'
 
 export class ValidationComposite implements Validation {
   private readonly validators: Validation[]
